+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- `#slice` - get a slice of config options as a hash set (works in a `#dig` manner);
+
 ## [0.8.0] - 2018-11-21
 ### Changed
 - `expose_yaml`, `load_from_yaml`, `load_from_json` and `load_from_self` treats empty hash (`{}`)